feat: add possibility for custom audit and report formatters
